# services/cache-service/api/openapi.yaml
openapi: 3.0.3
info:
  title: Cache Service API
  description: |
    High-performance Redis-based caching service for the microservices ecosystem.

    ## Features
    - **Sub-millisecond operations**: < 1ms GET, < 2ms SET operations
    - **High throughput**: 10,000+ operations/second sustained
    - **Ecosystem integration**: Profile, Task, and Session caching support
    - **Batch operations**: Efficient multi-key operations
    - **Circuit breaker patterns**: Resilience and fault tolerance
    - **Comprehensive monitoring**: Prometheus metrics and health checks

    ## Service Context
    This cache service supports the Profile-Service → Queue-Service → Worker-Service → Storage-Service ecosystem with specialized caching patterns for:
    - **Profile data**: User profiles, email lookups, metadata
    - **Task processing**: Task status, queue metrics, worker status
    - **Session management**: JWT blacklisting, session lifecycle

    ## Performance Targets
    - **Availability**: 99.9% uptime with proper failover
    - **Latency**: P99 < 5ms for all operations
    - **Error Rate**: < 1% under normal operations
    - **Cache Hit Ratio**: > 85% for optimal performance
  version: 1.0.0
  contact:
    name: Cache Service Team
    email: cache-service@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://cache-service.staging.example.com
    description: Staging environment
  - url: https://cache-service.prod.example.com
    description: Production environment

paths:
  # Health and Status Endpoints
  /health:
    get:
      summary: Health check endpoint
      description: Returns the health status of the cache service and Redis connectivity
      tags:
        - Health
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
        "503":
          description: Service is unhealthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /ready:
    get:
      summary: Readiness probe endpoint
      description: Returns readiness status for Kubernetes readiness probes
      tags:
        - Health
      responses:
        "200":
          description: Service is ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessResponse"
        "503":
          description: Service is not ready

  /metrics:
    get:
      summary: Prometheus metrics endpoint
      description: Returns Prometheus-format metrics for monitoring
      tags:
        - Monitoring
      responses:
        "200":
          description: Metrics data
          content:
            text/plain:
              schema:
                type: string
                example: |
                  # HELP cache_operations_total Total number of cache operations
                  # TYPE cache_operations_total counter
                  cache_operations_total{operation="get",result="hit"} 12345

  # Basic Cache Operations
  /api/v1/cache/{key}:
    get:
      summary: Get cached value
      description: Retrieve a value from the cache by key
      tags:
        - Cache Operations
      parameters:
        - name: key
          in: path
          required: true
          description: Cache key to retrieve
          schema:
            type: string
            maxLength: 512
            example: user:profile:12345
      responses:
        "200":
          description: Value found in cache
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CacheGetResponse"
        "404":
          description: Key not found in cache
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Invalid key format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      summary: Delete cached value
      description: Remove a key-value pair from the cache
      tags:
        - Cache Operations
      parameters:
        - name: key
          in: path
          required: true
          description: Cache key to delete
          schema:
            type: string
            maxLength: 512
            example: user:profile:12345
      responses:
        "200":
          description: Key deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CacheDeleteResponse"
        "404":
          description: Key not found
        "400":
          description: Invalid key format

  /api/v1/cache:
    post:
      summary: Set cached value
      description: Store a key-value pair in the cache with optional TTL
      tags:
        - Cache Operations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CacheSetRequest"
      responses:
        "200":
          description: Value stored successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CacheSetResponse"
        "400":
          description: Invalid request format or size limits exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Batch Operations
  /api/v1/cache/batch/get:
    post:
      summary: Batch get operation
      description: Retrieve multiple values from cache in a single operation
      tags:
        - Batch Operations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchGetRequest"
      responses:
        "200":
          description: Batch operation completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchGetResponse"
        "400":
          description: Invalid request or too many keys

  /api/v1/cache/batch/set:
    post:
      summary: Batch set operation
      description: Store multiple key-value pairs in a single operation
      tags:
        - Batch Operations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchSetRequest"
      responses:
        "200":
          description: Batch operation completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchSetResponse"
        "400":
          description: Invalid request or size limits exceeded

  /api/v1/cache/batch:
    delete:
      summary: Batch delete operation
      description: Delete multiple keys in a single operation
      tags:
        - Batch Operations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchDeleteRequest"
      responses:
        "200":
          description: Batch operation completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchDeleteResponse"

  # Pattern Operations
  /api/v1/cache/pattern/{pattern}:
    delete:
      summary: Delete by pattern
      description: Delete all keys matching a pattern (use with caution)
      tags:
        - Pattern Operations
      parameters:
        - name: pattern
          in: path
          required: true
          description: Pattern to match keys (supports wildcards)
          schema:
            type: string
            example: user:session:*
      responses:
        "200":
          description: Pattern deletion completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatternDeleteResponse"
        "400":
          description: Invalid pattern

  # TTL Management
  /api/v1/cache/ttl/{key}:
    get:
      summary: Get TTL for key
      description: Retrieve the time-to-live for a specific key
      tags:
        - TTL Management
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: TTL retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TTLResponse"
        "404":
          description: Key not found

    put:
      summary: Set TTL for key
      description: Update the time-to-live for a specific key
      tags:
        - TTL Management
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TTLRequest"
      responses:
        "200":
          description: TTL updated successfully
        "404":
          description: Key not found

  # Statistics and Monitoring
  /api/v1/cache/stats:
    get:
      summary: Cache statistics
      description: Get comprehensive cache performance statistics
      tags:
        - Statistics
      responses:
        "200":
          description: Statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CacheStatsResponse"

components:
  schemas:
    # Request/Response Models
    CacheSetRequest:
      type: object
      required:
        - key
        - value
      properties:
        key:
          type: string
          maxLength: 512
          description: Cache key
          example: user:profile:12345
        value:
          type: string
          maxLength: 1048576
          description: Value to store (JSON string or plain text)
          example: '{"id":12345,"name":"John Doe","email":"john@example.com"}'
        ttl:
          type: string
          description: Time-to-live duration (e.g., "1h", "30m", "300s")
          example: "1h"
          default: "1h"

    CacheSetResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Value stored successfully"
        key:
          type: string
          example: user:profile:12345
        expiration:
          type: string
          format: date-time
          example: "2024-01-01T13:00:00Z"

    CacheGetResponse:
      type: object
      properties:
        key:
          type: string
          example: user:profile:12345
        value:
          type: string
          example: '{"id":12345,"name":"John Doe","email":"john@example.com"}'
        ttl:
          type: integer
          description: Remaining TTL in seconds
          example: 3600
        cached_at:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00Z"

    CacheDeleteResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Key deleted successfully"
        key:
          type: string
          example: user:profile:12345

    # Batch Operation Models
    BatchGetRequest:
      type: object
      required:
        - keys
      properties:
        keys:
          type: array
          items:
            type: string
          maxItems: 100
          description: List of keys to retrieve
          example: ["user:profile:1", "user:profile:2", "user:profile:3"]

    BatchGetResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        results:
          type: object
          additionalProperties:
            type: string
          example:
            "user:profile:1": '{"id":1,"name":"Alice"}'
            "user:profile:2": '{"id":2,"name":"Bob"}'
        found:
          type: integer
          example: 2
        not_found:
          type: array
          items:
            type: string
          example: ["user:profile:3"]

    BatchSetRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CacheItem"
          maxItems: 100
        default_ttl:
          type: string
          example: "1h"

    BatchSetResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        processed:
          type: integer
          example: 3
        failed:
          type: array
          items:
            type: string
          example: []

    BatchDeleteRequest:
      type: object
      required:
        - keys
      properties:
        keys:
          type: array
          items:
            type: string
          maxItems: 100
          example: ["user:profile:1", "user:profile:2"]

    BatchDeleteResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        deleted:
          type: integer
          example: 2
        not_found:
          type: integer
          example: 0

    # TTL Models
    TTLRequest:
      type: object
      required:
        - ttl
      properties:
        ttl:
          type: string
          description: New TTL duration
          example: "30m"

    TTLResponse:
      type: object
      properties:
        key:
          type: string
          example: user:profile:12345
        ttl:
          type: integer
          description: TTL in seconds (-1 for no expiration, -2 for key not found)
          example: 1800
        expires_at:
          type: string
          format: date-time
          example: "2024-01-01T12:30:00Z"

    # Pattern Operations
    PatternDeleteResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        pattern:
          type: string
          example: "user:session:*"
        deleted:
          type: integer
          example: 15
        message:
          type: string
          example: "15 keys matching pattern deleted"

    # Statistics and Health
    CacheStatsResponse:
      type: object
      properties:
        cache_stats:
          type: object
          properties:
            total_keys:
              type: integer
              example: 12345
            memory_usage:
              type: string
              example: "128MB"
            hit_ratio:
              type: number
              format: float
              example: 0.85
            operations_per_second:
              type: number
              format: float
              example: 5432.1
        redis_stats:
          type: object
          properties:
            connected_clients:
              type: integer
              example: 25
            used_memory:
              type: string
              example: "256MB"
            uptime:
              type: string
              example: "5 days"
        performance_metrics:
          type: object
          properties:
            avg_latency_ms:
              type: number
              format: float
              example: 0.8
            p95_latency_ms:
              type: number
              format: float
              example: 2.1
            p99_latency_ms:
              type: number
              format: float
              example: 4.5

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          enum: [healthy, unhealthy]
          example: healthy
        timestamp:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00Z"
        version:
          type: string
          example: "1.0.0"
        checks:
          type: object
          properties:
            redis:
              type: object
              properties:
                status:
                  type: string
                  enum: [ok, error]
                  example: ok
                latency_ms:
                  type: number
                  format: float
                  example: 0.5
            circuit_breaker:
              type: object
              properties:
                state:
                  type: string
                  enum: [closed, open, half-open]
                  example: closed

    ReadinessResponse:
      type: object
      properties:
        ready:
          type: boolean
          example: true
        timestamp:
          type: string
          format: date-time

    # Helper Models
    CacheItem:
      type: object
      required:
        - key
        - value
      properties:
        key:
          type: string
          maxLength: 512
        value:
          type: string
          maxLength: 1048576
        ttl:
          type: string
          example: "1h"

    ErrorResponse:
      type: object
      properties:
        error:
          type: boolean
          example: true
        message:
          type: string
          example: "Key not found"
        code:
          type: string
          example: "KEY_NOT_FOUND"
        timestamp:
          type: string
          format: date-time
        details:
          type: object
          additionalProperties: true

  # Security (if authentication is required)
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service authentication

# Global security (uncomment if authentication is required)
# security:
#   - ApiKeyAuth: []

tags:
  - name: Health
    description: Health check and readiness endpoints
  - name: Cache Operations
    description: Basic cache CRUD operations
  - name: Batch Operations
    description: Efficient multi-key operations
  - name: Pattern Operations
    description: Pattern-based cache operations
  - name: TTL Management
    description: Time-to-live management
  - name: Statistics
    description: Cache performance statistics
  - name: Monitoring
    description: Monitoring and metrics endpoints